+++ android/src/main/kotlin/br/com/engapp/websocket_manager/models/ChannelName.kt
@@ -1,0 +1,9 @@
+package br.com.engapp.websocket_manager.models
+
+class ChannelName {
+    companion object {
+        const val PLUGIN_NAME = "websocket_manager"
+        const val MESSAGE = "websocket_manager/message"
+        const val DONE = "websocket_manager/done"
+    }
+}